refactor(harness): move built-in middleware impls from mod.rs into library/
middleware/mod.rs hosted five full middleware impls (Logging,
MessageTrim, ContextCompression, PromptCacheGuard, UsageAccounting)
alongside the module's own stack-runner wiring, while every other
built-in middleware already lived in library/. Moved
MessageTrim/ContextCompression/PromptCacheGuard into a new
library/context.rs and Logging/UsageAccounting into library/observe.rs
(they fit the existing observation concern). mod.rs now holds only
AgentRun helpers and the stack runner. No behavior change.
